+++ source/pio.h
@@ -62,9 +62,9 @@ extern void pio_init(void);
 extern void pio_reset(void);
 extern void pio_shutdown(void);
 extern void pio_ctrl_w(uint8_t data);
-extern uint8_t pio_port_r(uint32_t offset);
-extern void sio_w(uint32_t offset, uint8_t data);
-extern uint8_t sio_r(uint32_t offset);
-extern uint8_t coleco_pio_r(uint32_t port);
+extern uint8_t pio_port_r(int32_t offset);
+extern void sio_w(int32_t offset, int32_t data);
+extern uint8_t sio_r(int32_t offset);
+extern uint8_t coleco_pio_r(int32_t port);
 
-#endif /* PIO_H_ */
+#endif /* _PIO_H_ */
