+++ source/sound/fmintf.h
@@ -11,7 +11,7 @@ enum
 
 typedef struct 
 {
-    uint32_t latch;
+    uint8_t latch;
     uint8_t reg[0x40];
 } FM_Context;
 
@@ -19,12 +19,12 @@ typedef struct
 void FM_Init(void);
 void FM_Shutdown(void);
 void FM_Reset(void);
-void FM_Update(int16_t **buffer, uint32_t length);
-void FM_Write(uint32_t offset, uint32_t data);
+void FM_Update(int16_t **buffer, int32_t length);
+void FM_Write(uint32_t offset, uint8_t data);
 void FM_GetContext(uint8_t *data);
 void FM_SetContext(uint8_t *data);
 uint32_t FM_GetContextSize(void);
 uint8_t *FM_GetContextPtr(void);
-void FM_WriteReg(uint32_t reg, uint32_t data);
+void FM_WriteReg(uint8_t reg, uint8_t data);
 
 #endif /* FMINTF_H_ */
